mcp-faq/pyproject.toml
[project]
name = "fastmcp-faq-agent"
version = "0.1.0"
description = "Minimal FAQ agent with FastMCP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastmcp>=2.0.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.1",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
